Conflicts:
	graylog2-bootstrap/pom.xml
	graylog2-inputs/pom.xml
	graylog2-plugin-interfaces/pom.xml
	graylog2-radio/pom.xml
	graylog2-rest-client/pom.xml
	graylog2-rest-client/src/main/java/org/graylog2/restclient/lib/Version.java
	graylog2-rest-models/pom.xml
	graylog2-rest-routes/pom.xml
	graylog2-server/pom.xml
	graylog2-shared/pom.xml
	integration-tests/pom.xml
	pom.xml

+++ graylog2-rest-client/src/main/java/org/graylog2/restclient/models/api/responses/system/indices/DeflectorConfigResponse.java
@@ -34,7 +34,7 @@ public class DeflectorConfigResponse {
     public Period maxTimePerIndex;
 
     @JsonProperty("max_size_per_index")
-    public int maxSizePerIndex;
+    public long maxSizePerIndex;
 
     @JsonProperty("max_number_of_indices")
     public int maxNumberOfIndices;
